Refactored popup generation
Previous iteration was generating popups, and adding content to them but
not in a way that allowed for a lot of customization. It wasn't making
proper use of the PopupContent component. Also changed the click
listener to only listen for my selected layers, instead of querying the
layers on map click