Added base frame for project page
This is where projects will be displayed, in small cards with a
thumbnail and description. In a later update I'll add a modal with an
image carousel, click the thumbnail and it will open a modal containing
either a large view of the thumbnail, or a carousel of multiple images.